Ensure `getOwner(this)` is only rewritten in test context.
Previously, all usages of `getOwner(this)` were rewritten throughout the
entire file being transformed. This causes issues when _some_ usages of
`getOwner(this)` need to be left intact, while others need to be
rewritten.

This PR limits processing of `getOwner(this)` -> `this.owner` to only
the `test` callback functions root scope and all of the methods found in
the `options` hash. See fixtures for examples.